Update and rename Main_split_2.cpp to execution.cpp
