--- Main_split_2.cpp
@@ -1,49 +1,0 @@
-#include <iostream>
-#include<assert.h>
-#include "Main_split_1.h"
-
-using namespace TelCoColorCoder;
-
-void testNumberToPair(int pairNumber,
-    TelCoColorCoder::MajorColor expectedMajor,
-    TelCoColorCoder::MinorColor expectedMinor)
-{
-    TelCoColorCoder::ColorPair colorPair =
-    TelCoColorCoder::GetColorFromPairNumber(pairNumber);
-    std::cout << "Got pair " << colorPair.ToString() << std::endl;
-    assert(colorPair.getMajor() == expectedMajor);
-    assert(colorPair.getMinor() == expectedMinor);
-}
-
-void testPairToNumber(
-    TelCoColorCoder::MajorColor major,
-    TelCoColorCoder::MinorColor minor,
-    int expectedPairNumber)
-{
-    int pairNumber = TelCoColorCoder::GetPairNumberFromColor(major, minor);
-    std::cout << "Got pair number " << pairNumber << std::endl;
-    assert(pairNumber == expectedPairNumber);
-}
-
-void getManual() {
-    int pairNumber = 0;
- 
-    for (int i = 1; i <= 25; i++) {
-        pairNumber = i;
-        TelCoColorCoder::ColorPair manualcolorPair =
-            TelCoColorCoder::GetColorFromPairNumber(pairNumber);
-        std::cout << pairNumber <<"--->"<< manualcolorPair.ToString() << std::endl;
-
-    }
-}
-
-int main() {
-    testNumberToPair(4, TelCoColorCoder::WHITE, TelCoColorCoder::BROWN);
-    testNumberToPair(5, TelCoColorCoder::WHITE, TelCoColorCoder::SLATE);
-
-    testPairToNumber(TelCoColorCoder::BLACK, TelCoColorCoder::ORANGE, 12);
-    testPairToNumber(TelCoColorCoder::VIOLET, TelCoColorCoder::SLATE, 25);
-    getManual();
-
-    return 0;
-}
