+++ execution.cpp
@@ -1,0 +1,15 @@
+#include <iostream>
+#include<assert.h>
+#include "Main_split_2.h"
+
+
+int main() {
+    testNumberToPair(4, TelCoColorCoder::WHITE, TelCoColorCoder::BROWN);
+    testNumberToPair(5, TelCoColorCoder::WHITE, TelCoColorCoder::SLATE);
+
+    testPairToNumber(TelCoColorCoder::BLACK, TelCoColorCoder::ORANGE, 12);
+    testPairToNumber(TelCoColorCoder::VIOLET, TelCoColorCoder::SLATE, 25);
+    getManual();
+
+    return 0;
+}
